More meaningful error message on unsupported ARM-mode breakpoints
The GDB remote protocol reuses the 'kind' on ARM to determine
the breakpoint type. Although the previous message was still valid, it
could create confusion to the user, as the problem can be easily solved
by recompiling the code using the -mthumb flag. This new error message
should give a better hint to the user in such case.